Module sc_consensus_subspace::archiver

Consensus archiver module.

Contains implementation of archiving process in Subspace blockchain that converts blockchain history (blocks) into archived history (pieces).

Structs

SegmentHeadersStore

Persistent storage of segment headers

Functions

create_subspace_archiver

Crate an archiver task that will listen for importing blocks and archive blocks at K depth, producing pieces and segment headers (segment headers are then added back to the blockchain as store_segment_header extrinsic).

decode_block

Symmetrical to encode_block(), used to decode previously encoded blocks

encode_block

Encode block for archiving purposes

recreate_genesis_segment

Derive genesis segment on demand, returns Ok(None) in case genesis block was already pruned