Module sc_consensus_subspace::block_import

Block import for Subspace, which includes stateful verification and corresponding notifications.

In most cases block import happens after stateless block verification using verifier, the only exception to that is locally authored blocks.

Since verifier is stateless, the remaining checks in block import are those that require presence of the parent block or its state in the database. Specifically for Proof of Time individual checkpoints are assumed to be checked already and only PoT inputs need to be checked to correspond to the state of the parent block.

After all checks and right before importing the block notification (SubspaceLink::block_importing_notification_stream) will be sent that archiver among other things is subscribed to.

Structs

BlockImportingNotification

Notification with number of the block that is about to be imported and acknowledgement sender that can be used to pause block production if desired.

SubspaceBlockImport

A block-import handler for Subspace.

Enums

Error

Errors encountered by the Subspace authorship task.