Function sc_consensus_subspace::start_subspace_archiver

pub fn start_subspace_archiver<Block, Client>(
    subspace_link: &SubspaceLink<Block>, 
    client: Arc<Client>, 
    spawner: &impl SpawnEssentialNamed, 
    is_authoring_blocks: bool
) where
    Block: BlockT,
    Client: ProvideRuntimeApi<Block> + BlockBackend<Block> + HeaderBackend<Block> + Send + Sync + 'static,
    Client::Api: SubspaceApi<Block, FarmerPublicKey> + ObjectsApi<Block>, 

Start an archiver that will listen for imported blocks and archive blocks at K depth, producing pieces and root blocks (root blocks are then added back to the blockchain as store_root_block extrinsic).