Crate sc_consensus_subspace

Subspace Proof-of-Storage Consensus

Subspace is a slot-based block production mechanism which uses a Proof-of-Storage to randomly perform the slot allocation. On every slot, all the farmers evaluate their disk-based plot. If they have a tag (reflecting a commitment to a valid encoding) that it is lower than a given threshold (which is proportional to the total space pledged by the network) they may produce a new block. The proof of the Subspace function execution will be used by other peers to validate the legitimacy of the slot claim.

The engine is also responsible for collecting entropy on-chain which will be used to seed the given PoR (Proof-of-Replication) challenge. An epoch is a contiguous number of slots under which we will be using the same base PoR challenge. During an epoch all PoR outputs produced as a result of block production will be collected into an on-chain randomness pool. Epoch changes are announced one epoch in advance, i.e. when ending epoch N, we announce the parameters (i.e, new randomness) for epoch N+2.

Since the slot assignment is randomized, it is possible that a slot is claimed by multiple farmers, in which case we will have a temporary fork, or that a slot is not claimed by any farmer, in which case no block is produced. This means that block times are probabilistic.

The protocol has a parameter c [0, 1] for which 1 - c is the probability of a slot being empty. The choice of this parameter affects the security of the protocol relating to maximum tolerable network delays.

The fork choice rule is weight-based, where weight equals the number of primary blocks in the chain. We will pick the heaviest chain (more blocks) and will go with the longest one in case of a tie.

Modules

aux_schema

Schema for Subspace epoch changes in the aux-db.

notification

Utility module for handling Subspace client notifications.

Structs

ArchivedSegmentNotification

Archived segments notification with new pieces

Config

A slot duration. Create with get_or_compute.

Epoch

Subspace epoch information

NewSlotInfo

Information about new slot that just arrived

NewSlotNotification

New slot notification with slot information and sender for solution for the slot

SubspaceBlockImport

A block-import handler for Subspace.

SubspaceIntermediate

Intermediate value passed to block importer.

SubspaceLink

State that must be shared between the import queue and the authoring logic.

SubspaceParams

Parameters for Subspace.

SubspaceVerifier

A verifier for Subspace blocks.

SubspaceWorker

Worker for Subspace which implements Future<Output=()>. This must be polled.

SubspaceWorkerHandle

A handle to the Subspace worker for issuing requests.

Enums

Error

Errors encountered by the Subspace authorship task.

SubspaceRequest

Requests to the Subspace service.

Statics

INTERMEDIATE_KEY

Intermediate key for Subspace engine.

Functions

block_import

Produce a Subspace block-import object to be used later on in the construction of an import-queue.

find_pre_digest

Extract the Subspace pre digest from the given header. Pre-runtime digests are mandatory, the function will return Err if none is found.

import_queue

Start an import queue for the Subspace consensus algorithm.

start_subspace

Start the Subspace worker.

start_subspace_archiver

Start an archiver that will listen for imported blocks and archive blocks at K depth, producing pieces and root blocks (root blocks are then added back to the blockchain as store_root_block extrinsic).