Crate sc_consensus_subspace

sc-consensus-subspace is the core of Subspace consensus implementation.

You should familiarize yourself with Subnomicon and, ideally, protocol specifications. Documentation here assumes decent prior knowledge of the protocol on conceptual level and will not explain how the protocol works, it will instead explain how the protocol is implemented.

All of the modules here are crucial for consensus, open each module for specific details.

Modules

archiver

Consensus archiver responsible for archival of blockchain history, it is driven by block import pipeline.

aux_schema

Schema for Subspace block weight in the aux-db.

block_import

Block import for Subspace, which includes stateful verification and corresponding notifications.

notification

Utility module for handling Subspace client notifications.

slot_worker

Slot worker drives block and vote production based on slots produced in sc_proof_of_time.

verifier

Stateless and parallelized block verification that happens before block is imported (except for locally produced blocks that are imported directly).

Structs

SubspaceLink

State that must be shared between various consensus components.