Crate sc_consensus_subspace

Subspace Proof-of-Archival-Storage consensus

Subspace is a slot-based block production mechanism which uses a Proof-of-Archival-Storage to randomly perform the slot allocation. On every slot, all the farmers evaluate their disk-based plot. If they have a tag (reflecting a commitment to a valid encoding) that it is lower than a given threshold (which is proportional to the total space pledged by the network) they may produce a new block.

Core inputs to the Proof-of-Archival-Storage, such as global randomness and solution range come from the runtime, see pallet-subspace for details.

The fork choice rule is weight-based, where weight is derived from the distance between solution proposed in a block and the local challenge for particular farmer. The heaviest chain (represents a chain with more storage pledged to it) will be preferred over alternatives or longest chain is in case of a tie.

For a more in-depth analysis of Subspace consensus can be found in our consensus whitepaper.

This crate contains following major components:

worker (sc-consensus-slots) for claiming slots (block production)

block verifier that stateless verification of signature and Proof-of-Space

block import that verifies Proof-of-Archival-Storage and triggers archiving of the history

archiver worker triggered by block import that ensures history is archived and segment headers are produced at precisely the right time before finishing block import

License: GPL-3.0-or-later WITH Classpath-exception-2.0

Modules

aux_schema

Schema for Subspace block weight in the aux-db.

notification

Utility module for handling Subspace client notifications.

Structs

ArchivedSegmentNotification

Notification with block header hash that needs to be signed and sender for signature.

BlockImportingNotification

Notification with number of the block that is about to be imported and acknowledgement sender that can be used to pause block production if desired.

NewSlotInfo

Information about new slot that just arrived

NewSlotNotification

New slot notification with slot information and sender for solution for the slot.

RewardSigningNotification

Notification with a hash that needs to be signed to receive reward and sender for signature.

SegmentHeadersStore

Persistent storage of segment headers

SubspaceBlockImport

A block-import handler for Subspace.

SubspaceLink

State that must be shared between the import queue and the authoring logic.

SubspaceParams

Parameters for Subspace.

SubspaceSyncOracle

Subspace sync oracle that takes into account force authoring flag, allowing to bootstrap Subspace network from scratch due to our fork of Substrate where sync state of nodes depends on connected nodes (none of which will be synced initially).

SubspaceVerifier

A verifier for Subspace blocks.

SubspaceWorker

Worker for Subspace which implements Future<Output=()>. This must be polled.

Enums

Error

Errors encountered by the Subspace authorship task.

Functions

block_import

Produce a Subspace block-import object to be used later on in the construction of an import-queue.

create_subspace_archiver

Crate an archiver task that will listen for importing blocks and archive blocks at K depth, producing pieces and segment headers (segment headers are then added back to the blockchain as store_segment_header extrinsic).

get_chain_constants

Get chain constant configurations

import_queue

Start an import queue for the Subspace consensus algorithm.

slot_duration

Read configuration from the runtime state at current best block.

start_subspace

Start the Subspace worker.